requirements.txt
flask
youtube-transcript-api
transformers
python-docx
pdfkit